+++ src/BoostProjectData.ts
@@ -91,7 +91,7 @@ export class BoostProjectData implements IBoostProjectData {
             return;
         }
 
-        if (previous) {
+        if (previous && previous.sections) {
             sections = Object.keys(previous.sections);
             sections.forEach((section) => {
                 const sectionSummary = this.sectionSummary[section];
@@ -104,9 +104,8 @@ export class BoostProjectData implements IBoostProjectData {
                         previous.sections[section].errorCells;
                     sectionSummary.filesAnalyzed -= 1;
                 } else {
-                    boostLogging.error(
-                        `Section ${section} not found in sectionSummary`,
-                        false
+                    boostLogging.debug(
+                        `Updating BoostProjectData File Summaries: Previous ${section} sectionSummary not found`
                     );
                 }
             });
@@ -134,11 +133,10 @@ export class BoostProjectData implements IBoostProjectData {
                     sectionSummary.status = BoostProcessingStatus.notStarted;
                 }
             } else {
-                boostLogging.error(
-                    `Section ${section} not found in sectionSummary`,
-                    false
+                boostLogging.debug(
+                    `Updating BoostProjectData File Summaries: New File Summary ${section} sectionSummary not found`
                 );
-            }
+        }
         });
     }