+++ src/performance_function_controller.ts
@@ -6,6 +6,7 @@ import { BoostConfiguration } from './boostConfiguration';
 import * as vscode from 'vscode';
 import * as boostnb from './jupyter_notebook';
 import { fullPathFromSourceFile } from './extension';
+import { boostLogging } from './boostLogging';
 
 export const performanceFunctionKernelName = 'performance_function';
 export const performanceFunctionOutputType = 'performanceList';
@@ -70,8 +71,13 @@ export class BoostPerformanceFunctionKernel extends KernelControllerBase {
         const baseLineNumber = lineNumberBaseFromCell(cell);
 
         response.details.forEach((bug: any, index: number) => {
+            let calculatedLineNumber = baseLineNumber + bug.lineNumber;            
+            if (calculatedLineNumber < 1) {
+                calculatedLineNumber = 1;
+            }
+
             markdown += `${index + 1}. **Severity**: ${bug.severity}/10\n\n`;
-            markdown += `   **Line Number**: ${baseLineNumber + bug.lineNumber}\n\n`;
+            markdown += `   **Line Number**: ${calculatedLineNumber}\n\n`;
             markdown += `   **Bug Type**: ${bug.bugType}\n\n`;
             markdown += `   **Description**: ${bug.description}\n\n`;
             markdown += `   **Solution**: ${bug.solution}\n\n\n`;
@@ -102,9 +108,17 @@ export class BoostPerformanceFunctionKernel extends KernelControllerBase {
         const lineNumberBase = lineNumberBaseFromCell(cell);
         let diagnostics: vscode.Diagnostic[] = [];
         response.details.forEach((bug: any, index: number) => {
-            let range = new vscode.Range(lineNumberBase + bug.lineNumber - 1, 0, lineNumberBase + bug.lineNumber - 1, 0);
-            let diagnostic = new vscode.Diagnostic(range, `Severity: ${bug.severity}\n${bug.description}`, vscode.DiagnosticSeverity.Warning);
-            diagnostics.push(diagnostic);
+            response.details.forEach((bug: any, index: number) => {
+                let calculatedLineNumber = lineNumberBase + bug.lineNumber - 1;
+            
+                if (calculatedLineNumber < 1) {
+                    calculatedLineNumber = 1;
+                }
+            
+                let range = new vscode.Range(calculatedLineNumber, 0, calculatedLineNumber, 0);
+                let diagnostic = new vscode.Diagnostic(range, `Severity: ${bug.severity}\n${bug.description}`, vscode.DiagnosticSeverity.Warning);
+                diagnostics.push(diagnostic);
+            });
         });
         this._performanceIssueCollection.set(vscode.Uri.parse(sourceFile), diagnostics);
 
@@ -123,5 +137,10 @@ function lineNumberBaseFromCell(cell: vscode.NotebookCell | boostnb.BoostNoteboo
     }
 
     // Check if lineNumberBase is a number, if not, return 0
-    return typeof lineNumberBase === 'number' ? lineNumberBase : 0;
+    if (typeof lineNumberBase === 'number') {
+        return lineNumberBase;
+    }
+    else {
+        return 0;
+    }
 }